+++ sipx/rfc/__init__.py
@@ -1,0 +1,5 @@
+"""RFC reference data and constants.
+
+This package will contain SIP method registries, header field
+definitions, response code tables, and other RFC-derived constants.
+"""